+++ package.json
@@ -36,6 +36,7 @@
     "@electron-forge/plugin-vite": "^7.1.0",
     "@ledgerhq/logs": "^6.10.1",
     "@types/auto-launch": "^5.0.2",
+    "@types/debug": "^4.1.12",
     "@types/electron-localshortcut": "^3.1.0",
     "@types/react": "^18.2.38",
     "@types/react-dom": "^18.2.17",
